Hoist modal states to root with useContext
It is desirable to trigger the 'New Shop' modal visibility from multiple
places in the application. I've made use of useContext to store state
centrally.

Also refactored the NewShopActionSheet to optionally take a display
component. This is so I can trigger the action sheet with multiple
presentations – e.g. a button.