Use createEntityAdapter for shops reducer
Trying out `createEntityAdapter` as it provides standard CRUD hooks
leaving less room for error in CRUD logic.

Introducing `createEntityAdapter` caused the shape of the store to
change, which resulted in runtime errors from the `redux-persist`
library.

To solve the storage problem I hooked up a 'purge' action to the
persistor and the shops reducer with intent to purge the shops data.
Unfortunately, this purged all data from the local app (I'm still unsure
why). It did "fix" the problem with the storage layer however.